support multiple singing algorithms
Using a basic builder patter, Signers can be created that support
either RSA, DSA, or ECDSA.  A Signer instance can only support a
single key type.  The intended use is to just pass in a KeyPair, use
the defaults, and not need to worry about anything else.

Since a Signer now needs a KeyPair to be instantiated, the various
"get me a key" methods are moved to their own utility class.

Each signing algorithm also supports specifying a hash algorithm name
(ex: SHA256) and a string code representing the provider.  The
provider code is only used for the native RSA implementation (still
the default).  It is expected that these settings will be rarely used.

As part of this the recommended path to construct ThreadLocalSigner
has changed.  Previously it was easy to end up with multiple
ThreadLocalSigner instances.  This was mostly harmless (except for
resources cleanup) when everything was hard coded to be SHA256withRSA,
but quickly leads to errors when the signers have different
configuration.  It is now best to create a single ThreadLocalSigner
per key (ie usually just one) and pass that downstream.

This is formally a compatibility break because the public signature of
Signer has changed.  However, almost all of the changes were hard
coded constants (container "RSA"), or exposed internals.  No changes
(ie to java-manta) are needed to continue to only use RSA, with SHA256
and native acceleration.  Methods rendered obsolete or undesirable by
these change have been marked as deprecated and can be removed in the
next major release.

ref #9