allow NSS via PKS#11 to be used for ECDSA singing
On a reasonable public cloud container, SunPKCS11-NSS is roughly 50%
faster than BouncyCastle, making this reflection jungle gym
worthwhile.  If the reflection fails (lacking a logging framework) we
print to stderr and move on.

NOTE: This is bound to the internals of BouncyCastle, so trivial
upstream changes could break this class.

ref #9