support post 6.7 openssh fingerprints
Pre-6.7 OpenSSH used hex encoded md5 as a key fingerprint.  This was
changed to be the base64 encoded sha256 with a prefix (`SHA256`).
Triton and Manta only have server side support for the un-prefixed md5
fingerprint.

Following the approach of the node tooling in PUBAPI-1146, break apart
the key fingerprint specified by the user from what is sent to the
server.  Instead validate the user fingerprint, and calculate our own
fingerprint to send to the server.  In the long run this simplifies
the API (fewer things to pass around) and helps give better errors to
the user sooner (instead of waiting for an error from the server).

The details of how OpenSSH serializes fingerprints is based on the
implementation in node-sshpk.

ref #10