prefer modern java.time classes over unsafe legacy ones
Java 8 included JRS-310 which brought in modern date/time libraries.
The legacy java.util.Date is mutable(!), doesn't actually represent a
date(!!), and companion classes are not thead-safe.  In particular the
existing call to `SimpleDateFormat.format` was unsafe from multiple
threads, which left callers of Signer in perpetual risk of errors.

ref #34